+++ CHANGELOG.md
@@ -3,3 +3,12 @@
 *   Added GitHub link to the footer
 *   Added a sign up and login system
 *   Made the tab cloaker only overlay the tab instead of replace the page
+
+-   3/9/2023 12:00 PM
+
+*   Finished user system
+
+-   3/10/2023 8:30 AM
+
+*   Allowed accounts to download their localstorage and cookies
+*   Removed password from cookie data